% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.adjmat}
\alias{as.adjmat}
\alias{as.adjmat.network}
\alias{as.adjmat.list}
\alias{as.adjmat.formula}
\title{An alternative to \code{as.matrix} to retrieve adjacency matrix fast}
\usage{
as.adjmat(x)

\method{as.adjmat}{network}(x)

\method{as.adjmat}{list}(x)

\method{as.adjmat}{formula}(x)
}
\arguments{
\item{x}{An object to be coerced as an adjacency matrix.}
}
\description{
This function does not perform significant checks. Furthermore, this function
won't keep the row/col names.
}
