% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_adjmat}
\alias{as_adjmat}
\alias{as_adjmat.network}
\alias{as_adjmat.matrix}
\alias{as_adjmat.list}
\alias{as_adjmat.formula}
\title{An alternative to \code{as.matrix} to retrieve adjacency matrix fast}
\usage{
as_adjmat(x)

\method{as_adjmat}{network}(x)

\method{as_adjmat}{matrix}(x)

\method{as_adjmat}{list}(x)

\method{as_adjmat}{formula}(x)
}
\arguments{
\item{x}{An object to be coerced as an adjacency matrix.}
}
\description{
This function does not perform significant checks. Furthermore, this function
won't keep the row/col names.
}
