% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/count_stats.R
\name{count_available}
\alias{count_available}
\alias{count_stats}
\title{Count Network Statistics}
\usage{
count_available(i = 0L)

count_stats(X, terms, attrs = NULL)
}
\arguments{
\item{X}{List of square matrices. (networks)}

\item{terms}{Character vector with the names of the statistics to calculate.
Currently, the only available statistics are: '\Sexpr{paste(ergmito::count_available(), collapse="', '")}'.}
}
\value{
A matrix of size \code{length(X) * length(terms)} with the corresponding
counts of statistics.
}
\description{
This function is similar to what \link[ergm:summary_formula]{ergm::summary_formula} does, but it provides
a fast wrapper suited for matrix class objects.
}
\examples{
# DGP 
x <- powerset(5)
ans0 <- count_stats(x[1:20], c("mutual", "edges"))

# Calculating using summary_formula
fm <- x[[i]] ~ mutual + edges
ans1 <- lapply(1:20, function(i) {
  environment(fm) <- environment()
  ergm::summary_formula(fm)
})

ans1 <- do.call(rbind, ans1)

# Comparing
all.equal(unname(ans0), unname(ans1))

}
