% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{ergmito_gof}
\alias{ergmito_gof}
\alias{gof}
\alias{gof.ergmito}
\alias{print.ergmito_gof}
\alias{plot.ergmito_gof}
\title{Goodness of Fit diagnostics for ERGMito models}
\usage{
gof(...)

\method{gof}{ergmito}(object, probs = c(0.05, 0.95), R = 50000L, ...)

\method{print}{ergmito_gof}(x, ...)

\method{plot}{ergmito_gof}(x, y = NULL,
  main = "Goodness-of-fit Statistics", ...)
}
\arguments{
\item{...}{Further arguments passed to \link[stats:quantile]{stats::quantile}.}

\item{object}{An object of class \link{ergmito}.}

\item{probs}{Numeric vector. Quantiles to plot (passed to \link[stats:quantile]{stats::quantile}).}

\item{R}{Integer scalar. Number of simulations to generate (passed to \link{sample}).}

\item{x}{An object of class \code{ergmito_gof}.}

\item{y}{Ignored.}

\item{main}{Title of the plot}
}
\value{
An object of class \code{ergmito_gof}. This is a list with the following
components:
\itemize{
\item \code{ci} A list of matrices of length \code{nnets(object)} with the corresponding
confidence intervals for the statistics of the model.
\item \code{target.stats} A matrix of the target statistics.
\item \code{ergmito.probs} A list of numeric vectors of length \code{nnets(object)} with the
probabilities associated to each possible structure of network.
\item \code{probs} The value passed via \code{probs}.
\item \code{model} The fitted model.
\item \code{term.names} Character vector. Names of the terms used in the model.
\item \code{quantile.args} A list of the values passed via \code{...}.
}
}
\description{
Goodness of Fit diagnostics for ERGMito models
}
\details{
The Goodness of Fit function uses the fitted ERGMito to calculate a given confidence
interval for a set of sufficient statistics. By default (and currently the
only available option), this is done on the sufficient statistics specified
in the model.

In detail, the algorithm is executed as follow:

For every network in the list of networks do:
\enumerate{
\item Calculate the probability of observing each possible graph in its support
using the fitted model.
\item Draw \code{R} samples from each set of parameters using the probabilities computed
in (1).
\item Using the \code{quantile} function, calculate the desired quantiles of the
sufficient statistics.
}

The plot method is particularly convenient since it graphically shows whether
the target statistics of the model (observed statistics) fall within the
simulated range.

The print method tries to copy (explicitly) the print method of the
\code{gof} function from the \code{ergm} R package.
}
\examples{
# Fitting the fivenets model
data(fivenets, package = "ergmito")
fit <- ergmito(fivenets ~ edges + nodematch("female"))

# Calculating the gof
ans <- gof(fit)

# Looking at the results
ans
plot(ans)
}
