% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fivenets}
\alias{fivenets}
\title{Example of a group of small networks}
\format{An object of class \code{list} of length 5.}
\usage{
fivenets
}
\description{
This list of networks was generated using the \link{new_rergmito} sampler from a
set of 5 baseline networks with a random vector of \code{female}. The sufficient
statistics that generate this data are \code{edges} and  \code{nodematch("female")} with parameters
-2.0 and 2.0 respectively.
}
\keyword{datasets}
