% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{lergm_boot}
\alias{lergm_boot}
\alias{lergm_boot.formula}
\alias{lergm_boot.lergm}
\alias{print.lergm_boot}
\title{Bootstrap of lergm}
\usage{
lergm_boot(x, ..., R, ncpus = 1L, cl = NULL)

\method{lergm_boot}{formula}(x, ..., R, ncpus = 1L, cl = NULL)

\method{lergm_boot}{lergm}(x, ..., R, ncpus = 1L, cl = NULL)

\method{print}{lergm_boot}(x, ...)
}
\arguments{
\item{x}{Either a formula or an object of class \link{lergm}.}

\item{...}{Additional arguments passed to the method.}

\item{R}{Integer. Number of replicates}

\item{ncpus}{Integer Number of CPUs to use.}

\item{cl}{An object of class \code{cluster} (see \link[parallel:makePSOCKcluster]{parallel::makePSOCKcluster})}
}
\description{
Bootstrap of lergm
}
\examples{

# Simulating 20 bernoulli networks of size 4
nets <- replicate(20, rbernoulli(4), simplify = FALSE)
}
