% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{lergm_formulae}
\alias{lergm_formulae}
\title{Processing formulas in \code{lergm}}
\usage{
lergm_formulae(model, stats = NULL, obs_stats = NULL,
  env = parent.frame(), ...)
}
\arguments{
\item{model}{A formula. The left-hand-side can be either a small network, or
a list of networks.}

\item{stats}{A list.}

\item{obs_stats}{Observed statistics. If multiple networks, then a list, otherwise
a named vector (see \link[ergm:summary_formula]{ergm::summary_formula}).}

\item{env}{Environment in which \code{model} should be evaluated.}

\item{...}{Further arguments passed to \link[ergm:ergm.allstats]{ergm::ergm.allstats}.}
}
\value{
A list of class \code{lergm_loglik}.
\itemize{
\item \code{loglik} A function. The log-likelihood function.
\item \code{grad} A function. The gradient of the model.
\item \code{stats} If the number of networks is greater than 1, then a list of objects
as returned by \link[ergm:ergm.allstats]{ergm::ergm.allstats} (one per network). Otherwise, a single
list returned by the same function.
\item \code{model} A formula. The model passed.
\item \code{npars} Integer. Number of parameters.
\item \code{nnets} Integer. Number of networks to estiamte.
}
}
\description{
The little ERGMs R package allows estimating pulled ERGMs by aggregating
independent networks together.
}
