% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{new_ergmito_ptr}
\alias{new_ergmito_ptr}
\title{Creates a new \code{ergmito_ptr}}
\usage{
new_ergmito_ptr(target_stats, stats_weights, stats_statmat)
}
\arguments{
\item{target_stats, stats_weights, stats_statmat}{see \link{exact_loglik}.}
}
\description{
After calculating the support of the sufficient statisitcs, the second
most computationally expensive task is computing log-likelihoods,
gradients, and hessians or ERGMs. This function creates a pointer to an
underlying class that is optimized to improve memory allocation and
save computation time when possible.
}
\details{
This function is for internal used only. Non-advanced users
are not encouraged to use it. See \link{ergmito_formulae} and \link{exact_loglik}
for user friendly wrappers of this function.
}
