% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{new_rergmito}
\alias{new_rergmito}
\alias{[.ergmito_sampler}
\alias{print.ergmito_sampler}
\title{ERGMito sampler}
\usage{
new_rergmito(model, theta = NULL, sizes = NULL, mc.cores = 2L, ...)

\method{[}{ergmito_sampler}(x, i, j, ...)

\method{print}{ergmito_sampler}(x, ...)
}
\arguments{
\item{model}{A formula.}

\item{theta}{Named vector. Model parameters.}

\item{sizes}{Integer vector. Values between 2 to 5 (6 becomes too intensive).}

\item{mc.cores}{Integer. Passed to \link[parallel:mclapply]{parallel::mclapply}}

\item{...}{Further arguments passed to \link[ergm:ergm.allstats]{ergm::ergm.allstats}.}

\item{x}{An object of class \code{ergmito_sampler}.}

\item{i, j}{\code{i} is an integer vector indicating the indexes of the networks to
draw, while \code{j} the corresponding sizes. These need not to be of the same size.}
}
\value{
An environment with the following objects:
\itemize{
\item \code{calc_prob}
\item \code{call}
\item \code{counts}
\item \code{networks}
\item \code{prob}
\item \code{sample} A function to draw samples. \code{n} specifies the number of samples to
draw, \code{s} the size of the networks, and \code{theta} the parameter to use to
calculate the likelihoods.
\item \code{theta}
}

The indexing method \code{[.ergmito_sampler} returns a named list of length
\code{length(j)}.
}
\description{
Using
}
\details{
The indexing method, \code{[.ergmito_sampler}, allows extracting networks
directly by passing indexes. \code{i} indicates the index of the networks to draw,
which go from 1 through \code{2^(n*(n-1))}, and \code{j} indicates the requested
size.
}
