% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_rergmito.R
\name{new_rergmito2}
\alias{new_rergmito2}
\title{Generate an ergmito model}
\usage{
new_rergmito2(model, theta, ...)
}
\arguments{
\item{theta}{passed}

\item{...}{Passed to \code{\link[=ergmito_formulae]{ergmito_formulae()}}.}

\item{object}{Either a model or an object of class \link{ergmito}.}
}
\description{
Generate an ergmito model
}
