% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{new_rlergm}
\alias{new_rlergm}
\alias{print.lergm_sampler}
\title{Little ERGM sampler}
\usage{
new_rlergm(model, theta = NULL, sizes = 2:4, mc.cores = 2L, ...)

\method{print}{lergm_sampler}(x, ...)
}
\arguments{
\item{model}{A formula.}

\item{theta}{Named vector. Model parameters.}

\item{sizes}{Integer vector. Values between 2 to 5 (6 becomes too intensive).}

\item{mc.cores}{Integer. Passed to \link[parallel:mclapply]{parallel::mclapply}}

\item{...}{Further arguments passed to \link[ergm:ergm.allstats]{ergm::ergm.allstats}.}

\item{x}{An object of class \code{lergm_sampler}.}
}
\description{
Using
}
