% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{rbernoulli}
\alias{rbernoulli}
\title{Random bernoulli graph}
\usage{
rbernoulli(n, p = 0.5)
}
\arguments{
\item{n}{Integer. Size of the graph}

\item{p}{Probability of a tie}
}
\value{
A square matrix of size \code{n} with zeros in the diagonal.
}
\description{
Random bernoulli graph
}
\examples{
# A graph of size 4
rbernoulli(4)
}
