% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_adjmat}
\alias{as_adjmat}
\title{An alternative to \code{as.matrix} to retrieve adjacency matrix fast}
\usage{
as_adjmat(x)
}
\arguments{
\item{x}{An object to be coerced as an adjacency matrix.}
}
\description{
This function does not perform significant checks. Furthermore, this function
won't keep the row/col names.
}
\examples{

set.seed(1231)
x <- matrix_to_network(rbernoulli(rep(5, 100)))
benchmarkito(
  as_adjmat = as_adjmat(x),
  as.matrix = lapply(x, as.matrix)
)
}
