% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm_blockdiagonal.R
\name{blockdiagonalize}
\alias{blockdiagonalize}
\alias{splitnetwork}
\title{Block-diagonal models using \code{ergm}}
\usage{
blockdiagonalize(x, attrname = "block")

splitnetwork(x, attrname)
}
\arguments{
\item{x}{In the case of \code{blockdiagonalize}, a list of networks or matrices.
For \code{splitnetwork} a single network object with a vertex attribute that can
be used to split the data.}

\item{attrname}{Name of the attribute that holds the block ids.}
}
\description{
These two functions are used to go back and forth from a pooled ergm vs a
blockdiagonal model, the latter to be fitted using \link[ergm:ergm]{ergm::ergm}.
}
\examples{
\dontrun{
library(ergm)
data(fivenets)

fivenets2 <- blockdiagonalize(fivenets, attrname = "block") # A network with
ans0 <- ergm(
  fivenets2 ~ edges + nodematch("female"),
  constraints = ~blockdiag("block")
  )
ans1 <- ergmito(fivenets ~ edges + nodematch("female"))

}
}
