% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergmito.R
\name{ergmito}
\alias{ergmito}
\alias{print.ergmito}
\alias{summary.ergmito}
\alias{coef.ergmito}
\alias{logLik.ergmito}
\alias{vcov.ergmito}
\title{Estimation of ERGMs using Exact likelihood functions}
\usage{
ergmito(model, stats = NULL, optim.args = list(method = "L-BFGS-B",
  lower = -100, upper = 100), ...)

\method{print}{ergmito}(x, ...)

\method{summary}{ergmito}(object, ...)

\method{coef}{ergmito}(object, ...)

\method{logLik}{ergmito}(object, ...)

\method{vcov}{ergmito}(object, ...)
}
\arguments{
\item{model}{Model to estimate. See \link[ergm:ergm]{ergm::ergm}. The only difference with
\code{ergm} is that the LHS can be a list of networks.}

\item{stats}{List as returned by \link[ergm:ergm.allstats]{ergm::ergm.allstats}. When this is provided,
the function does not call \code{ergm.allstats}, which can be useful in simulations.}

\item{optim.args}{List. Passed to \link[stats:optim]{stats::optim}.}

\item{...}{Further arguments passed to the method. In the case of \code{ergmito},
\code{...} are passed to \link{ergmito_formulae}.}

\item{x, object}{An object of class \code{ergmito}}
}
\value{
An list of class \code{ergmito}:
\itemize{
\item \code{coef}          Named vector. Parameter estimates.
\item \code{iterations}    Integer. Number of times the loglikelihood was evaluated
(see \link[stats:optim]{stats::optim}).
\item \code{loglikelihood} Numeric. Final value of the objective function.
\item \code{covar}         Square matrix of size \code{length(coef)}. Variance-covariance matrix
\item \code{coef.init}     Named vector of length \code{length(coef)}. Initial set of parameters
used in the optimization.
\item \code{formulae}      An object of class \link[=ergmito_formulae]{ergmito_loglik}.
\item \code{network}       Networks passed via \code{model}.
}
}
\description{
As a difference from \link[=ergm]{ergm::ergm}, \code{ergmito} uses the exact log-likelihood
function for fitting the model. This implies that all the \code{2^(n*(n-1))}
graphs are generated for computing the normalizing constant of the ERGM
model. This implies that models with up to 5 nodes are relatively simple
to fit, since more than that can become infeasible.
}
\examples{

# Generating a small graph
set.seed(12)
n <- 4
net <- sna::rgraph(n, tprob = .7)

model <- net ~ edges + mutual + balance

library(ergm)
ans_ergmito <- ergmito(model)
ans_ergm  <- ergm(model)

# The ergmito should have a larger value
ergm.exact(ans_ergmito$coef, model)
ergm.exact(ans_ergm$coef, model)

summary(ans_ergmito)
summary(ans_ergm)
}
\seealso{
The function \link{plot.ergmito} for post-estimation diagnostics.
}
