% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{exact_loglik}
\alias{exact_loglik}
\alias{exact_gradient}
\title{Vectorized calculation of ERGM exact loglikelihood}
\usage{
exact_loglik(x, params, stats.weights, stats.statmat)

exact_gradient(x, params, stats.weights, stats.statmat)
}
\arguments{
\item{x}{Matrix. Observed statistics}

\item{params}{Numeric vector. Parameter values of the model.}

\item{stats.weights, stats.statmat}{Vector and Matrix as returned by \link[ergm:ergm.allstats]{ergm::ergm.allstats}.}
}
\description{
This function can be compared to \link[ergm:ergm.exact]{ergm::ergm.exact} with the statistics
centered at \code{x}, the observed statistic.
}
