% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fivesamplers}
\alias{fivesamplers}
\title{Five ERGMito samplers}
\format{An object of class \code{list} of length 5.}
\usage{
fivesamplers
}
\description{
This list contains five ERGMito samplers. Each one of these was built using
a random Bernoulli graph with an attribute \code{female}. The parameters used for
creating the sampler are \verb{(edges = -2.0, nodematch("female") = 2.0)}. A example
of a dataset generated with this is \link{fivenets}.
}
\keyword{datasets}
