% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{new_rergmito}
\alias{new_rergmito}
\alias{print.ergmito_sampler}
\title{ERGMito sampler}
\usage{
new_rergmito(model, theta = NULL, sizes = 2:4, mc.cores = 2L, ...)

\method{print}{ergmito_sampler}(x, ...)
}
\arguments{
\item{model}{A formula.}

\item{theta}{Named vector. Model parameters.}

\item{sizes}{Integer vector. Values between 2 to 5 (6 becomes too intensive).}

\item{mc.cores}{Integer. Passed to \link[parallel:mclapply]{parallel::mclapply}}

\item{...}{Further arguments passed to \link[ergm:ergm.allstats]{ergm::ergm.allstats}.}

\item{x}{An object of class \code{ergmito_sampler}.}
}
\value{
An environment with the following objects:
\itemize{
\item \code{calc_prob}
\item \code{call}
\item \code{counts}
\item \code{networks}
\item \code{prob}
\item \code{sample} A function to draw samples. \code{n} specifies the number of samples to
draw, \code{s} the size of the networks, and \code{theta} the parameter to use to
calculate the likelihoods.
\item \code{theta}
}
}
\description{
Using
}
