% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nvertex}
\alias{nvertex}
\alias{nedges}
\alias{nedges.network}
\alias{nedges.list}
\alias{nedges.matrix}
\alias{nedges.ergmito}
\alias{nedges.formula}
\alias{nvertex.network}
\alias{nvertex.matrix}
\alias{nvertex.list}
\alias{nvertex.ergmito}
\alias{nvertex.formula}
\alias{nnets}
\alias{nnets.list}
\alias{nnets.matrix}
\alias{nnets.network}
\alias{nnets.ergmito}
\alias{nnets.formula}
\title{Utility functions to query network dimensions}
\usage{
nvertex(x)

nedges(x, ...)

\method{nedges}{network}(x, ...)

\method{nedges}{list}(x, ...)

\method{nedges}{matrix}(x, ...)

\method{nedges}{ergmito}(x, ...)

\method{nedges}{formula}(x, ...)

\method{nvertex}{network}(x)

\method{nvertex}{matrix}(x)

\method{nvertex}{list}(x)

\method{nvertex}{ergmito}(x)

\method{nvertex}{formula}(x)

nnets(x)

\method{nnets}{list}(x)

\method{nnets}{matrix}(x)

\method{nnets}{network}(x)

\method{nnets}{ergmito}(x)

\method{nnets}{formula}(x)
}
\arguments{
\item{x}{Either an object of class \link{ergmito}, \link{network}, \link{formula}, or \link{matrix}.}

\item{...}{Further arguments passed to the method. Currently only \code{nedges.network}
receives arguments (see \link[network:network.edgecount]{network::network.edgecount}).}
}
\description{
Utility functions to query network dimensions
}
