% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nvertex}
\alias{nvertex}
\alias{nvertex.network}
\alias{nvertex.matrix}
\alias{nvertex.list}
\alias{nvertex.lergm}
\alias{nnets}
\alias{nnets.list}
\alias{nnets.lergm}
\title{Utility functions to query network dimensions}
\usage{
nvertex(x)

\method{nvertex}{network}(x)

\method{nvertex}{matrix}(x)

\method{nvertex}{list}(x)

\method{nvertex}{lergm}(x)

nnets(x)

\method{nnets}{list}(x)

\method{nnets}{lergm}(x)
}
\arguments{
\item{x}{Either an object of class \link{lergm}, \link{network}, or \link{matrix}.}
}
\description{
Utility functions to query network dimensions
}
