% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nvertex}
\alias{nvertex}
\alias{nvertex.network}
\alias{nvertex.matrix}
\alias{nvertex.list}
\alias{nvertex.ergmito}
\alias{nvertex.formula}
\alias{nnets}
\alias{nnets.list}
\alias{nnets.matrix}
\alias{nnets.network}
\alias{nnets.ergmito}
\title{Utility functions to query network dimensions}
\usage{
nvertex(x)

\method{nvertex}{network}(x)

\method{nvertex}{matrix}(x)

\method{nvertex}{list}(x)

\method{nvertex}{ergmito}(x)

\method{nvertex}{formula}(x)

nnets(x)

\method{nnets}{list}(x)

\method{nnets}{matrix}(x)

\method{nnets}{network}(x)

\method{nnets}{ergmito}(x)
}
\arguments{
\item{x}{Either an object of class \link{ergmito}, \link{network}, or \link{matrix}.}
}
\description{
Utility functions to query network dimensions
}
