% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{rbernoulli}
\alias{rbernoulli}
\title{Random bernoulli graph}
\usage{
rbernoulli(n, p = 0.5)
}
\arguments{
\item{n}{Integer vector. Size of the graph. If \code{length(n) > 1}, then it will
a list of random graphs.}

\item{p}{Probability of a tie}
}
\value{
A square matrix of size \code{n} with zeros in the diagonal.
}
\description{
Random bernoulli graph
}
\examples{
# A graph of size 4
rbernoulli(4)

# 3 graphs of various sizes
rbernoulli(c(3, 4, 2))
}
